requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
PyPDF2
chromadb
faiss-cpu
langchain-google-genai
ipython